Refactor settings by putting the settings closer to the actual document
