+++ src/DependencyInjection/SetonoSyliusAlgoliaExtension.php
@@ -8,7 +8,6 @@ use Setono\SyliusAlgoliaPlugin\DataMapper\DataMapperInterface;
 use Setono\SyliusAlgoliaPlugin\Filter\Doctrine\FilterInterface as DoctrineFilterInterface;
 use Setono\SyliusAlgoliaPlugin\Filter\Object\FilterInterface as ObjectFilterInterface;
 use Setono\SyliusAlgoliaPlugin\Provider\IndexScope\IndexScopeProviderInterface;
-use Setono\SyliusAlgoliaPlugin\Provider\IndexSettings\IndexSettingsProviderInterface;
 use Setono\SyliusAlgoliaPlugin\UrlGenerator\ResourceUrlGeneratorInterface;
 use Symfony\Component\Config\FileLocator;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
@@ -60,9 +59,6 @@ final class SetonoSyliusAlgoliaExtension extends Extension
 
         $container->registerForAutoconfiguration(IndexScopeProviderInterface::class)
             ->addTag('setono_sylius_algolia.index_scope_provider');
-
-        $container->registerForAutoconfiguration(IndexSettingsProviderInterface::class)
-            ->addTag('setono_sylius_algolia.index_settings_provider');
 
         $container->registerForAutoconfiguration(ObjectFilterInterface::class)
             ->addTag('setono_sylius_algolia.object_filter');
