+++ src/Resources/config/services/indexer.xml
@@ -7,7 +7,7 @@
             <argument type="service" id="doctrine"/>
             <argument type="service" id="setono_sylius_algolia.provider.index_scope.composite"/>
             <argument type="service" id="setono_sylius_algolia.index_name_resolver.default"/>
-            <argument type="service" id="setono_sylius_algolia.registry.index_settings_provider"/>
+            <argument type="service" id="setono_sylius_algolia.provider.index_settings"/>
             <argument type="service" id="setono_sylius_algolia.data_mapper.composite"/>
             <argument type="service" id="setono_sylius_algolia.command_bus"/>
             <argument type="service" id="serializer"/>
